Solves redeclaration of background and background type
